New Azeem feedbacks
Login Page: Check the mockup I sent on Monday. Add a second div that shows the 3 listed features with emoji (📈 ✅ 🔭). Also, below the header SurveySight logo, add the action words "Monitor. Track. Analyze."

Help/FAQ on Footer: The "?" emoji looks like a partially loaded component. Instead, write in bold <b>Frequently Asked Questions</b> and put a darker background for it to form a button.

Top-right Corner Icon: In ISAS, we used to display logged in user's name. In this SurveySight, next to the icon , add "Hi, <First Name>"

Full Screen Mode - Refresh Icon: When in full screen mode, hover mouse on the refresh icon, change "Go Fullscreen" to "Screen Refresh Timer"

Full Screen Mode - Timer Size & Opacity:  When in full screen mode, make the timer size a little smaller. Also, maintain the current opacity ONLY when a user hovers mouse on it. If the mouse is not in the region of the timer, reduce the opacity to make it more transparent/faint (so that it is not blocking charts/numbers)